refactor(CartPresenter): CartProduct 리스트 생성을 데이터 모듈로 분리
